refactor: move test-utils to template root
